requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
matplotlib>=3.7.0
openpyxl>=3.1.0
pyyaml>=6.0
scikit-learn>=1.3.0
pydantic>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0
